Put Donations link back in prominent location.
This link is too important for the project to let it disappear in a part of the site that our users almost never look at, as the past has sadly proven.
I have no problem if people say they don't want to donate because hey why pay a little bit for something if you can get it free. But i don't want to hear something like "I didn't know one could donate to it". Having Donations directly below "Download" - the only link people tend to click, is exactly the right spot.